+++ gestionecorsi/src/it/betacom/businesscomponent/idgenerator/IdGeneratorCorso.java
@@ -29,14 +29,14 @@ public class IdGeneratorCorso implements IdGeneratorInterface, DAOConstants {
 	}
 
 	@Override
-	public long nextId() throws DAOException, ClassNotFoundException, IOException {
-		long id = 0;
+	public int nextId() throws DAOException, ClassNotFoundException, IOException {
+		int id = 0;
 
 		try {
 			stmt = conn.createStatement();
 			rs = stmt.executeQuery(SELECT_CORSOSEQ);
 			rs.next();
-			id = rs.getLong(1);
+			id = rs.getInt(1);
 		} catch (SQLException e) {
 			throw new DAOException(e);
 		}
