+++ gestionecorsi/src/test/it/betacom/businesscomponent/idgenerator/IdGeneratorCorsistaTest.java
@@ -17,7 +17,7 @@ class IdGeneratorCorsistaTest {
 			
 			IdGeneratorCorsista idGen = IdGeneratorCorsista.getIstance();			
 			assertNotNull(idGen, "Istanza create correttamente");			
-			long valore = idGen.nextId();			
+			int valore = idGen.nextId();			
 			assertEquals(valore, idGen.nextId() - 1, "Test sequenza");			
 		} catch (DAOException | ClassNotFoundException | IOException e) {
 			e.printStackTrace();
